requirements.txt
brotlipy==0.7.0
certifi==2021.10.8
click==8.0.4
docker-pycreds==0.4.0
filelock==3.6.0
gitdb==4.0.9
GitPython==3.1.27
huggingface-hub==0.4.0
joblib==1.1.0
mkl-fft==1.3.1
mkl-service==2.4.0
numpy==1.22.3
packaging==21.3
pandas==1.4.1
pathtools==0.1.2
Pillow==9.0.1
promise==2.3
protobuf==3.19.4
psutil==5.9.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
regex==2022.3.15
sacremoses==0.0.49
sentry-sdk==1.5.8
setproctitle==1.2.2
shortuuid==1.0.8
smmap==5.0.0
termcolor==1.1.0
tokenizers==0.11.6
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.63.0
transformers==4.17.0
wandb==0.12.11
yaspin==2.1.0
